Add rpm build stage
Standard bdist_rpm won't work out of the gate. The generated spec file
calls a post script that calls the /usr/lib/rpm/brp-python-bytecompile
script to bytecompile all found python files. This breaks because of
the pkg.py jinja template because it's not a syntactically valid python
file. Not only does this script have to be neutralized to side-step this
issue, but the spec file also has to be patched so that the install
phase doesn't attempt to bytecompile the file either. The install phase
records all files created to the INSTALLED_FILES file which is fed into
the spec file's %files macro, causing it to search for bytecompiled
.pyc and .pyo files that don't exist because of the neutralized script.
Only together does the old-school way of generating an RPM file via the
rpmbuild command succeed.

I'm not crazy about this dirty hack, but the saving grace is that by only
relying on bdist_rpm to generate a starting point spec file, it becomes
possible to further patch the spec file with gitlab CI context data like
tag names etc to work into the RPM packaging process.